requirements.txt
Flask>=3.0.0
flask-session>=0.4.0
python-dotenv>=1.0.1
stripe>=11.6.0